+++ src/util.h
@@ -12,7 +12,7 @@ char * get_tox_ID(Tox * tox);
 
 void bootstrap(Tox * tox);
 
-void get_elapsed_time_str(char *buf, int bufsize, time_t secs);
+void get_elapsed_time_str(char *buf, size_t bufsize, time_t secs);
 
 bool file_exists(const char *filename);
